separate read from http and fs
