+++ app/vmalert/config/fs.go
@@ -2,8 +2,6 @@ package config
 
 import (
 	"fmt"
-	"io"
-	"net/http"
 	"net/url"
 	"strings"
 	"sync"
@@ -34,39 +32,17 @@ var (
 	fsRegistry   = make(map[string]FS)
 )
 
-// readFromFSOrHTTP reads path either from filesystem or from http if path starts with http or https.
-// when reading from filesystem, parses the given path list and inits FS for each item.
-// Once initialed, readFromFSOrHTTP will try to read and return files from each FS.
-// readFromFSOrHTTP returns an error if at least one FS failed to init.
+// readFromFS parses the given path list and inits FS for each item.
+// Once initialed, readFromFS will try to read and return files from each FS.
+// readFromFS returns an error if at least one FS failed to init.
 // The function can be called multiple times but each unique path
 // will be initialed only once.
 //
-// It is allowed to mix different FS types and url in path list.
-func readFromFSOrHTTP(paths []string) (map[string][]byte, error) {
+// It is allowed to mix different FS types in path list.
+func readFromFS(paths []string) (map[string][]byte, error) {
+	var err error
 	result := make(map[string][]byte)
 	for _, path := range paths {
-		if isHTTPURL(path) {
-			// reads remote file via http or https, if url is given
-			resp, err := http.Get(path)
-			if err != nil {
-				return nil, fmt.Errorf("cannot fetch %q: %w", path, err)
-			}
-			data, err := io.ReadAll(resp.Body)
-			_ = resp.Body.Close()
-			if resp.StatusCode != http.StatusOK {
-				if len(data) > 4*1024 {
-					data = data[:4*1024]
-				}
-				return nil, fmt.Errorf("unexpected status code when fetching %q: %d, expecting %d; response: %q", path, resp.StatusCode, http.StatusOK, data)
-			}
-			if err != nil {
-				return nil, fmt.Errorf("cannot read %q: %s", path, err)
-			}
-			result[path] = data
-			continue
-		}
-
-		var err error
 		fsRegistryMu.Lock()
 		fs, ok := fsRegistry[path]
 		if !ok {
@@ -108,7 +84,6 @@ func readFromFSOrHTTP(paths []string) (map[string][]byte, error) {
 			result[k] = v
 		}
 	}
-
 	return result, nil
 }
